requirements.txt
python-telegram-bot==20.7
playwright==1.40.0
aiofiles==23.2.1
jinja2==3.1.2
pytz>=2024.1
flask==3.0.0
requests==2.31.0

